# protocol/openapi.yaml
openapi: 3.0.3
info:
  title: Remote File Explorer — Host Agent API
  version: 1.17.0
  description: >
    Contract between the Flutter mobile app and the Go host agent. JSON for control and
    metadata; binary streams for file content. All routes except /pair, /register,
    /login, /auth/challenge, and /health require a bearer device token. Clients may
    send an optional `X-RFE-Client-Version` header (e.g.
    "1.10.0+18") on every authenticated request; the agent records this plus the
    caller's network address against the device row. Admin devices see those fields
    for all devices via GET /devices (see Device.lastAddress / Device.lastVersion);
    ordinary paired devices receive only their own minimal device record.

    A device's token is "admin" (Device.viaLogin=true) if it was obtained via /login or
    /register (proof of the single account's password) rather than /pair (a one-time
    code meant for an ordinary device like the phone app). Admin devices may manage
    OTHER devices: POST /pairing/generate, and PATCH/DELETE /devices/{id} targeting a
    device other than themselves. An ordinary (non-admin) device may still manage only
    ITSELF via DELETE /devices/{id}.

    Pair, register, and login all require device-identity proof-of-possession: fetch a
    nonce from POST /auth/challenge, sign it with the device's permanent Ed25519 key,
    and send the (devicePublicKey, nonce, signature) triple. The agent pins the public
...
servers:
...
    variables:
...
security:
  - deviceToken: []
paths:
...
  /settings:
    get:
      summary: Get agent settings
      description: >
        Admin devices (`isAdmin=true`) receive host-wide settings. Ordinary
        paired devices receive `effectiveScope=device`, with roots and readOnly
        reflecting the policy applied to that device's request. The response
        omits unrelated host-wide settings; `agentName` and `allowSharing` are
        intentionally shared so clients can identify the host and show the
        share-link policy. It returns the photo-backup
        destination to a paired device only when that device's effective roots
        allow the path and its effective policy permits writes; this is a
        policy check and does not guarantee the operating system will permit
        the eventual write. Otherwise the configured/available booleans report
        the state without disclosing the destination.
      security: [{ deviceToken: [] }]
      responses:
        '200':
          description: Current settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentSettings' }
        '401': { $ref: '#/components/responses/Error' }
    patch:
      summary: Update host-wide agent settings (admin-only, partial)
      description: Requires an admin device authenticated through /login or /register.
      security: [{ deviceToken: [] }]
      requestBody:
        content:
          application/json:
            schema: { $ref: '#/components/schemas/AgentSettings' }
      responses:
        '200':
          description: Updated settings
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AgentSettings' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '403':
          description: Caller is not an admin device.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Error' }
...
components:
  securitySchemes:
    deviceToken:
      type: http
      scheme: bearer
  responses:
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
  schemas:
...
    AgentSettings:
      type: object
      properties:
        isAdmin:
          type: boolean
          description: True when the caller may manage host-wide settings and other devices.
        effectiveScope:
          type: string
          enum: [global, device]
          description: Global host policy for admins, or this device's effective policy for paired clients.
        accessDenied:
          type: boolean
          description: >
            For paired clients, true when a per-device jail falls outside all
            configured host roots. In that state an empty roots list means no
            filesystem access, not unrestricted access. False for admins.
        readOnly: { type: boolean }
        roots:
          type: array
          items: { type: string }
        agentName: { type: string }
        allowSharing:
          type: boolean
          description: >
            Host-level gate for R1 one-time share links (default false).
            POST /share/mint returns 403 while this is false.
        photoBackupConfigured:
          type: boolean
          description: Whether a photo-backup destination is configured; safe to return to all devices.
        photoBackupAvailable:
          type: boolean
          description: >
            Whether the effective path roots and read-only policy permit this
            caller to write to the configured photo-backup destination. This
            does not guarantee operating-system filesystem permissions. For
            paired clients, photoBackupRoot is returned only when this is true.
        photoBackupRoot:
          type: string
          description: >
            PC-side destination folder for phone photo backup, set from the
            web companion Settings page only. Returned to admins and to paired
            clients only when photoBackupAvailable is true. Empty means no
            destination was disclosed.
...
    Error:
      type: object
      description: >
        Uniform error envelope returned on non-2xx responses. The `code` field is
        the authoritative machine-readable vocabulary the agent emits (see enum);
        `message` is a human-readable detail string and may change without notice.
      properties:
        code:
          type: string
          example: PATH_NOT_FOUND
          enum:
            - BAD_REQUEST
            - UNAUTHORIZED
            - READ_ONLY
            - FORBIDDEN
            - PATH_NOT_FOUND
            - CONFLICT
            - CHUNK_HASH_MISMATCH
            - HASH_MISMATCH
            - NOT_AVAILABLE
            - THUMB_BUSY
            - INVALID_CODE
            - DELETE_FAILED
            - RATE_LIMITED
            - RESOURCE_LIMIT
            - TRANSFER_ACTIVE
            - TRANSFER_NOT_OPEN
            - PAYLOAD_TOO_LARGE
            - NOT_FOUND
            - INTERNAL
          description: >
            - `BAD_REQUEST` (400): malformed or missing required fields.
            - `UNAUTHORIZED` (401): missing/invalid/revoked bearer token.
            - `READ_ONLY` (403): write rejected — the agent is in global
              read-only mode OR the calling device has the per-device read-only
              flag set (#8).
            - `FORBIDDEN` (403): path escapes the configured root jail; or a
              /devices/{id} DELETE/PATCH or /pairing/generate call that
              requires an admin device (Device.viaLogin=true) came from a
              non-admin device.
            - `PATH_NOT_FOUND` (404): path does not exist.
            - `NOT_FOUND` (404): PATCH /devices/{id} — no device with that id.
            - `CONFLICT` (409): destination already exists.
            - `CHUNK_HASH_MISMATCH` (409): per-chunk SHA-256 did not match
              X-Chunk-Sha256.
            - `HASH_MISMATCH` (422): whole-file SHA-256 did not match on complete.
            - `NOT_AVAILABLE` (404): no thumbnail/preview available for this file.
            - `THUMB_BUSY` (429): thumbnail request capacity is full; retry after the `Retry-After` interval.
            - `INVALID_CODE` (401): pairing code invalid or expired.
            - `DELETE_FAILED`: per-item batch-delete failure (returned inside a
              BatchResult entry, not as a top-level HTTP error).
            - `RATE_LIMITED` (429): too many /pair attempts (10/min).
            - `RESOURCE_LIMIT` (429): upload admission reached the per-device
              (4 sessions / 64 GiB) or host-wide (16 sessions / 128 GiB)
              session or reserved-byte quota. Retry after completing or
              deleting an existing session; no fixed Retry-After is provided.
            - `TRANSFER_ACTIVE` (409): an operation conflicts with active
              session work; completion is exclusive, and DELETE is rejected
              while another request holds the session.
            - `TRANSFER_NOT_OPEN` (409): the session is completed or failed
              and no longer accepts chunks or completion.
            - `PAYLOAD_TOO_LARGE` (413): chunk body exceeds the session's chunkSize.
            - `INTERNAL` (500): unexpected server error.
        message: { type: string }
        details: { type: object, additionalProperties: true }